# packages/contracts/openapi/openapi.yaml
openapi: 3.1.1
info:
  title: SaaS Template API
  version: 1.0.0
  description: >-
    Language-neutral HTTP API contract for the SaaS platform foundation. Major
    API versions are encoded in the server base path.
servers:
  - url: /api/v1
    description: Version 1 API
tags:
  - name: Authentication
  - name: Users
  - name: Organizations
  - name: Workspaces
paths:
  /auth/register:
    post:
      operationId: register
      summary: Create an owner account, organization, and opaque session
      tags: [Authentication]
      security: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account and session established
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Set-Cookie:
              description: HttpOnly opaque session cookie; attributes are controlled by the API.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      operationId: login
      summary: Establish an opaque server-side session
      tags: [Authentication]
      security: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Session established
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Set-Cookie:
              description: Opaque session cookie; attributes are controlled by the API.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      operationId: logout
      summary: Revoke the current session
      tags: [Authentication]
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '204':
          description: Session revoked and the session cookie expired
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Set-Cookie:
              description: Expired HttpOnly session cookie.
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me:
    get:
      operationId: getCurrentUser
      summary: Return the current authenticated user
      tags: [Users]
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Current user
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/current:
    get:
      operationId: getCurrentOrganization
      summary: Return the organization selected for the request
      tags: [Organizations]
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationId'
      responses:
        '200':
          description: Current organization
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/current:
    get:
      operationId: getCurrentWorkspace
      summary: Return the workspace selected for the request
      tags: [Workspaces]
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Current workspace
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: >-
        Opaque HttpOnly session credential managed by the API. Secure and
        SameSite attributes are deployment-configured; only its hash is stored.
  parameters:
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Optional caller-provided correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    OrganizationId:
      name: X-Organization-Id
      in: header
      required: true
      description: Organization selected for the request.
      schema:
        type: string
        format: uuid
    WorkspaceId:
      name: X-Workspace-Id
      in: header
      required: true
      description: Workspace selected for the request.
      schema:
        type: string
        format: uuid
  headers:
    RequestId:
      description: Correlation identifier assigned or accepted by the API.
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is required or invalid
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The current principal lacks permission for the resource
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The selected resource does not exist
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: The requested identity conflicts with an existing record
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 1024
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 1024
    AuthResponse:
      type: object
      additionalProperties: false
      required: [user, session]
      properties:
        user:
          $ref: '#/components/schemas/CurrentUserResponse'
        session:
          type: object
          additionalProperties: false
          required: [expiresAt]
          properties:
            expiresAt:
              type: string
              format: date-time
    CurrentUserResponse:
      $ref: '#/components/schemas/User'
    User:
      type: object
      additionalProperties: false
      required: [id, email, displayName, emailVerifiedAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        emailVerifiedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Organization:
      type: object
      additionalProperties: false
      required: [id, name, slug, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Workspace:
      type: object
      additionalProperties: false
      required: [id, organizationId, name, slug, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Membership:
      type: object
      additionalProperties: false
      required:
        [id, userId, organizationId, workspaceId, roleIds, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        workspaceId:
          type: [string, 'null']
          format: uuid
        roleIds:
          type: array
          items:
            type: string
            format: uuid
          uniqueItems: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Role:
      type: object
      additionalProperties: false
      required:
        [
          id,
          organizationId,
          key,
          name,
          isSystem,
          permissions,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        description:
          type: [string, 'null']
        isSystem:
          type: boolean
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Permission:
      type: object
      additionalProperties: false
      required: [id, key, description, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        description:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Session:
      type: object
      additionalProperties: false
      required: [id, userId, expiresAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            details:
              type: object
              additionalProperties: true
    PaginationMeta:
      type: object
      additionalProperties: false
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0